fix: fix code for proper picture validation
